+++ @navikt/core/react/src/datepicker/rdp/DatePickerRDPInternal.tsx
@@ -7,10 +7,10 @@ import {
 } from "react-day-picker";
 import { Show } from "../../primitives/responsive";
 import { omit } from "../../utils-external";
+import { getLocaleFromString } from "../../utils/components/date/Date.locale";
+import { isDateRange } from "../../utils/components/date/Date.typeutils";
+import { useDateInputContext } from "../../utils/components/date/input/DateInput";
 import { clampDisplayMonth, isDateOutsideRange } from "../../utils/date";
-import { useDateInputContext } from "../../utils/date/Date.Input";
-import { getLocaleFromString } from "../../utils/date/Date.locale";
-import { isDateRange } from "../../utils/date/Date.typeutils";
 import { pickRangeSelection } from "../../utils/date/pick-range-selection";
 import { cl } from "../../utils/helpers";
 import { useDateLocale } from "../../utils/i18n/i18n.hooks";
