+++ @navikt/core/react/src/datepicker/root/useRangeDatepicker.tsx
@@ -1,10 +1,10 @@
 import { differenceInCalendarDays, isSameDay, isWeekend } from "date-fns";
 import React, { useState } from "react";
 import { dateMatchModifiers } from "react-day-picker";
+import { getLocaleFromString } from "../../utils/components/date/Date.locale";
+import type { DateRange } from "../../utils/components/date/Date.typeutils";
+import type { DateInputProps } from "../../utils/components/date/input/DateInput";
 import { formatDateForInput, isValidDate, parseDate } from "../../utils/date";
-import type { DateInputProps } from "../../utils/date/Date.Input";
-import { getLocaleFromString } from "../../utils/date/Date.locale";
-import type { DateRange } from "../../utils/date/Date.typeutils";
 import { focusElement } from "../../utils/helpers/focus";
 import { useDateLocale } from "../../utils/i18n/i18n.hooks";
 import type { DatePickerProps } from "./DatePickerRoot";
