+++ @navikt/core/react/src/modal/Modal.tsx
@@ -4,7 +4,7 @@ import { createPortal } from "react-dom";
 import { useProvider } from "../provider/Provider";
 import { Detail, Heading } from "../typography";
 import { useId } from "../utils-external";
-import { useDateInputContext } from "../utils/date/Date.Input";
+import { useDateInputContext } from "../utils/components/date/input/DateInput";
 import { cl, composeEventHandlers } from "../utils/helpers";
 import { useMergeRefs, useScrollLock } from "../utils/hooks";
 import { ModalContextProvider, useModalContext } from "./Modal.context";
