+++ @navikt/core/react/src/monthpicker/root/MonthPicker.util.ts
@@ -1,6 +1,6 @@
 import { setYear } from "date-fns";
+import type { Matcher } from "../../utils/components/date/Date.typeutils";
 import { isMatch } from "../../utils/date";
-import type { Matcher } from "../../utils/date/Date.typeutils";
 
 export const nextEnabled = (
   months: Date[],
