+++ @navikt/core/react/src/utils/components/date/Date.typeutils.ts
