+++ @navikt/core/react/src/utils/date/is-match.ts
@@ -10,7 +10,7 @@ import {
   isDateAfterType,
   isDateBeforeType,
   isDateRange,
-} from "./Date.typeutils";
+} from "../components/date/Date.typeutils";
 
 function isDateType(value: unknown): value is Date {
   return isDate(value);
